one more DecimalField(default=0, decimal_places=2, max_digits=10)

+++ app_run/models.py
@@ -13,7 +13,7 @@ class Run(models.Model):
     comment = models.TextField()
     athlete = models.ForeignKey(User, on_delete=models.CASCADE)
     status = models.CharField(max_length=30, choices=STATUS_CHOICES, default='init')
-    distance = models.FloatField(default=0)
+    distance = models.DecimalField(default=0, decimal_places=2, max_digits=10)
     speed = models.DecimalField(default=0, decimal_places=2, max_digits=10)
     run_time_seconds = models.IntegerField(default=0)
